UART

+++ Source/02-ECUA/LM35/Inc/Lm35.h
@@ -9,7 +9,7 @@
 #define SOURCE_02_ECUA_LM35_INC_LM35_H_
 
 void Lm35_vidInit(void);
-u8 Lm35_u8TempRead(void);
+u16 Lm35_u8TempRead(void);
 
 
 #endif /* SOURCE_02_ECUA_LM35_INC_LM35_H_ */
